rpc: Add address family to unresolved_address
Add optional address family to rpc::unresolved_address. If the field
affects name resolution. If it's not provided the name can be resolved
into ipv4 or ipv6 address (current behavior). The parameter could be
used to specify a desired resolution outcome.

This commit also sets ipv4 as a resolution target for archival storage.