c/backend: handling moving not initialized partition cross shards
When partition is being moved cross shards, it is shut down on
source core and recreated on target one. X-core move doesn't involve
any data movement. When partition is recreated on target core we relay
on data in partition underlying log to recover partition state. The bug
that was triggered in ducktape tests was caused by moving partition
before any data were persisted to disk. This prevented partition on
target core from recovering its initial configuration.

Fixed the problem by initializing partition on target core with its
original configuration. This way if there is no data it will start from
valid initial state, or if data are in log this initial configuration
will be overridden.

Signed-off-by: Michal Maslanka <michal@vectorized.io>