rpk container start: add --set to override config values
Recently, a `--set` flag was added to `rpk redpanda start`, which
allows setting config values at startup. This changeset introduces
a similar `--set` flag to `rpk container start`, but instead of
parsing the values, both the flags and the values are passed through
to the containers' entrypoint (`rpk redpanda start`).