test: add verifiable consumer from upstream kafka
Adds the verifiable consumer from the upstream kafka ducktape test
suite. This wraps the verifiable consumer that is part of the standard
kafka java cli tools.

This verifiable consumer prints metadata to standard out which is parsed
by the wrapper in order to write tests that assert properties about what
has been consumed.

Signed-off-by: Noah Watkins <noah@vectorized.io>