s/readers_cache: disposing invalid readers when returned to cache
Fixed bug leading to situation in which caller have to wait for
operation requiring log write log until reader where evicted from cache
by eviction timer.

The problem was related with the fact that when `cached_reader`
`entry_guard` was destroyed it didn't scheduled disposing readers that
were marked as invalid but only those which were no longer reusable.
This might lead to the situation in which reader that was in use during
truncate call might not be disposed immediately after it was destroyed
even tho it was marked as invalid by `readers_cache::evict_if` method.
The reader was reinserted to the readers cache, preventing truncation to
finish. The reader was then removed in the next eviction timer
iteration.

Example leading to this bug occurrence:

1) create reader
2) truncate log in a range for which reader holds a lock
3) destroy a reader
4) truncation should finish immediately after reader was destroyed

Fixes: #1426

Signed-off-by: Michal Maslanka <michal@vectorized.io>