tx coordinator: fix transaction atomicity issue
Redpanda keeps ongoing transactions in memory to avoid write latency.
When a re-election happens we lose this information. Since the Kafka's
protocol initiates a transaction implicitly on the first request then
when a request lends on new leader it starts new transaction and a
client commits only the second part assuming the whole transaction.

We fix this problem by setting a fencing (with ready status) when a
transaction starts on a new leader and by failing a transaction if
current the leader's term doesn't match the fence. When this happens
a client aborts a transaction and the abort refreshes the fence.