tx coordinator: use lww conflict resolution
Supporting deletion clean policy forced us to use last write wins (lww)
conflict resolution instead of first write wins.

Why do conflicts exist in the first place?

Upon a request a transaction coordinator uses its in-memory state to
make a decision, writes a command (the decision) to its partition, waits
until a catch up processes reads the command back, replays it to update
the in memory state and replies to the client.

When a write fails the transaction coordinator doesn't know if the write
has passed or not so it doesn't wait for the catch up process and times
out the request. Because the status of the write is unknown the state of
the transaction coordinator may become stale. In that case the next
command may conflict with the previous one but since the client doesn't
get an acknowledgement the coordinator may choose any command to resolve
the conflict.

We used to do first write wins conflict resolution but it doesn't work
well with deletion clean up policy. Let's look on the following change
log where the conflicts are designated by the same letter:

    a, b, c1, c2, d, e

With first write wins policy `c1` takes precedence over `c2` but when we
cut the beginning of the log just after `c1`:

    c2, d, e

`c2` becomes first and we resolve the conflict in its favor potentially
violating consistency. Using last write wins (lww) policy eliminates
this scenario.