test: add partition movement test with dymamic io
Adds an end-to-end test that runs a producer and consumer in the
background while partitions are being moved and verifying the process
after the movement is done.

This test is currently disabled due to some failure. For no movement or
a few movements the test passes but with more aggresive amounts of
partition movements an offset error occurs.

Signed-off-by: Noah Watkins <noah@vectorized.io>