archival: Add 'remote' component
The component is responsible for uploading and downloading data to/from
S3. It can handle manifests and segments.

The remote component implements backoff and error handling in a way that
makes sense for archival storage. It deals with S3 specific things like
SlowDown responses and "short read" errors from http client. It also
uses manifests to locate the segments in S3.

Apply suggestions from code review

Co-authored-by: Ben Pope <BenPope@users.noreply.github.com>