rpk topic set-config: Allow updating the config incrementally
`rpk topic set-config` uses the AlterConfig Kafka API RPC, which
replaces all config values with the defaults and sets the new values.

The IncrementalAlterConfig RPC isn't supported by sarama, so this
change is a workaround that firsts fetches the current config, sets
the new values, and sends all the config fields, achieving similar
results, but from the client side.