rpk: add support for plugins
This follows the exact algorithm that is used in kubernetes for plugins;
a followup PR will introduce a `plugin list` subcommand that will
behavior similarly to kubernetes as well.

Plugin support is fairly simple; the test is a bit overkill.

Unlike kubernetes, I'd like to bundle in plugin search directly into
rpk, rather than eschew responsibility to a separate command (krew).
That can also be done as followup.