tx coordinator: distinguish user and auto abort
Kafka protocol uses implicit transaction start: first add_partition_to_tx
or add_offsets_to_tx RPC of a transaction starts a transaction. When it
encounters a previous transaction in the aborting state it re-aborts it
and start a new transaction. If the auto aborted transactions ends up in
the aborting state it may lead to a problem:

  * a user initiate a transaction (via add_partition_to_tx)
  * the user hangs
  * txn coordinator auto aborts the transaction
  * the user wakes up and send another add_partition_to_tx
  * it re-aborts the transaction and implicitly starts a new one
  * user commits the transaction

As a result user thinks that it comitted two add_partition_to_tx but in
fact it comitted only one.

This commit introduced a new state `killed` to distinguish user aborts
and auto aborts.