storage: store segment appender as pointer
Prior to this patch the segment stored its segment appender as a value.
This is generally fine, but when closing the segment the appender is
moved around by value. If there are no on-going operations with the
segment (which may have captured a reference to `this`) then this is
fine.

However, the reason this has not been problematic is that existing cases
of moving the segment appender have been proceeded by calls to the
appenders flush() method which waited on all operations to complete.

This is overly restrictive of a method like flush whose purpose should
be only guarantee that data is on disk--inserting a global barrier is
too expensive. In a later patch we relax this rule for flush() to
support concurrent flushes and appends. However, we will no longer able
to rely on flush to ensure that there are no longer any on going
operations.

Existing uses of the appender do correctly call close() which does
insert the barrier during shutdown, and to deal with the issue of moving
the appender following a flush in the code path for rolling a segment
we've changed the segment to store the appender on the heap instead of a
value so that captured reference remain valid.

Signed-off-by: Noah Watkins <noah@vectorized.io>