storage: fix flushing performance regression
Prior to this change when writes completed they would examine if any
flushes should be dispatched based on the set of pending flush ops.
Once the flush was complete, any ops that fell below the newly flushed
offset would be completed.

The regression was that when many writes were inflight, many writes
would make the same determination that a flush was necessary. This
resulted in far more flushes than necessary.

The fix was to _remove_ those ops to be completed by the flush into
flush's continuation so that the next write to complete would not
observe those pending flush operations being handled by the other fiber.

Signed-off-by: Noah Watkins <noah@vectorized.io>