task/dev: add task for installing golang
Adds a task for installing golang, taking into account version numbers.
The task is parametrized so that a GOLANG_URL can be given to install Go
from a URL other than the official site.

Golang is installed to `{{.BUILD_ROOT}}/go/`, with a subfolder created
for each distinct version. A GO_VERSION variable is defined, as well as
GO_BUILD_ROOT one that points to the right folder.

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>