tx coordinator: add barrier in addition to sync
persisted_stm::sync guatatees that a state machine has catched up with
all the commands replicated in previous term but the state maybe stale
(undecided) if a replication failed within current term. It may be fine
for read-modify-write requests (lww policy) but doesn't work for the
read only requests, in this case we need to write a barrier to be sure
the state is up to date.