tx coordinator: don't overwrite killed with aborting
It's important not to misuse `killed` status as `aborting`. With the
latter we expect a transacional request to re-abort the last transaction
and silently start a new one. If we do the same with `killed` it will
lead to a partially commited transaction. We need to fail the request
and for a client to refresh its session.