c/ctrl_backend: fixed bootstrapping controller backend with x-core moves
When partition is moved cross cores `controller_backend` uses currently
existing partition instance as a source for revision that the moved
partition have to be created with. When bootstrapping the partition on
source node isn't available. We have to establish revision of partition
that it have to be created with during bootstrap process.

Added code that finds operation which revision was used to create
the partition.

Signed-off-by: Michal Maslanka <michal@vectorized.io>