(split) k8s/dockerfile: make dockerfile a template
The Dockerfile is a bit different depending on whether is consumed by
`docker build` or `docker buildx`. In order to accommodate both cases,
this commit makes the existing Dockerfile a "template" that need to be
pre-processed when regular `docker build` is used; for buildx/buildkit
the file can be used "as is".

Signed-off-by: Ivo Jimenez <ivo@vectorized.io>