security: filter prefix acls using resource type
When searching the stored ACLs for resource name prefixes the resource
type was not being taken into account. This allowed, for example, a
prefix rule on a topic to authorize an operation on a group when the
group had a name and the topic was a prefix of the name.

Fixes: #1579

Signed-off-by: Noah Watkins <noah@vectorized.io>