rpk: add autocompletion to plugins
This is a bit less straightforward than just simple rpk autocompletion.

Rather than only figure out what plugins are available when we see args
that rpk does not understand, we can list all plugins always on startup.
These can be used for autocompletion.

The way that autocompletion works within cobra is that all commands and
their subcommands get dedicated functions within the generated shell
scripts. So, to get autocompletion of our listed plugins, we need our
plugins to be listed in the shell scripts, and to do that, we have to
add a whole bunch of tiny cobra.Commands corresponding to our plugins
before rpk executes the "generate shell-autocomplete" command.

Lastly, as one extra point of improvement, this defines "rpk.ac-" to be
the opt-in prefix to use if a plugin supports the --help-autocomplete
flag. If a plugin supports this, then it can return the use, long,
short, and valid args that we can use in our tiny cobra.Commands to make
them actually useful when running `rpk foo bar --help`. Basically, we
get the help text from the plugin, and we get its autocomplete args.

I've tested this all locally with a few scripts printing hello, and one
Go program that supports --help-autocomplete that returns

    ext
    ext-foo
    ext-foo-bat
    ext-bar