ducktape: Fix async rpk consumer
Update the command to use the new "topic" namespace.

Sometimes, the consumer might start before the cluster is ready (e.g.
it's in the middle of an election), so add a retry loop to account
for that.

Override clean_node to prevent warning messages from being printed
each time the consumer is started.