rpk start: Log an error when an unknown flag is passed
Unknown flags are allowed in `rpk start` because not all of them
can be parsed by cobra (pflag) specifically, JSON literals to
override the configuration, since they usually contain many commas
and pflag interprets that as a CSV.

However, this also means that no errors are logged if the user
makes a typo, which could lead to much confusion. This changeset
adds a function to check all flags, exiting the command with an
error if one of them isn't recognized.