rpk: override viper's default SupportedExts to avoid picking up non-yaml
viper does not use our explicit `SetConfigType("yaml")` when it searches
for configs with ReadInConfig. Instead, it uses SupportedExts. Since
that variable is global, we can just override it to be our list of (1)
extension that we want to search for.

This is a small fix, the real one is that we will be removing viper from
rpk soon enough.

I've manually tested the worklow that exists in #1637, we no longer have
mentions of trying to use json.

Fixes #1637.