s3: Recompute derived keys
AWS signature requires computation of the derived key which includes
the date component. This derived key expires after two days. Current
implementation assumes that the client is created per-request. This
is no longer the case after client pool was introduced to s3.

This commit adds ability to check the expiration date and recompute
the derived key if needed. The method should be called manually
before signing the header.