archival: Improve observability in case of error
Archival subsystem has the upload loop (UL) which is a fiber
that runs asynchronously for every ntp. It's responsibe for
uploading new segments and maintaining manifests. In case of
successful upload it starts upload next set of segments immediately.

If there is nothing to upload the loop uses exponential backoff
limiting max backoff interval to 5s. It start with pretty short
backoff interval of 10ms. This was done to prevent high CPU utilization
if there is no data to upload.

In case of error it's more complicated. The retriable errors (like
SlowDown response from S3) gets retried and backoff is applied
per request. This reties respect the total time budget of the
upload. But if the error is not retrieable (for instance NoSuchKey)
no retries are performed and the error is propagated.

After any error upload loop goes to the next iteration, picks the same
set of segments to upload and starts again. This leads to pretty
busy error output if the error is not retrieable.

This commit makes upload loop use exponential backoff if no progress
was made on the previous iteration.

Also, this commit discards the result of the failed future in the
reconciliation loop. The failure was handled but exceptional future
wasn't handled correctly which lead to warning being added to the log.