c/scheduling: refactored partition allocator to evaluate constraints
This commit introduces changes in partition allocation process to make
it easy to pass separate constraints set for each allocated partitions.
In this commit partition allocation state was separated from the
allocator performing the allocation additionally it introduces the
allocation strategy which may be implemented using sophisticated
constraint solver. Partition allocation is an atomic process and
allocator guarantee that internal state is correct in face of allocation
failures.

Signed-off-by: Michal Maslanka <michal@vectorized.io>