tx coordinator: expire old transactions
Transaction coordinator used to keep meta info about all ever existed
transactions in memory (and in its log). With time it could lead to a
resource leak and OOM errors.

Fixing it by writing tombstones and removing all information about the
expired transactions.

The transaction coordinator's GC process uses the last update time and
current system time to determine when it should delete the expired tran-
sactions.

The last update time may be recorded on a different node compared to the
current one so it is subject to clock skew.

It doesn't affect safety but may temporarily affect availability when
a re-election happens and a new leader is far-far away in the future.
In this case all ongoing transactions will be aborted but the follow up
transactions won't be affected.

By default transactions expire in 7 days so it's fine if the time on the
nodes drift up to several hours.