archival: Fix leadership trasfer upload bug
After leadership transfer the new leader might issue a partial
upload (since the segments are not aligned across nodes). This can
lead to lack of global upload progress.

The bug is caused by the partial upload imprecision. It can start
uploading new segment not from the offset that is required but from
the nearby offset which is smaller than requred (since the index
is used for loockup).

This commit also adds the reproducer for the problem.