coproc: Read from cluster::partition
- In all instances where log->make_reader is called replace with
partition->make_reader

- Remove all instances where log->offsets().dirty_offset is used and replaced
with partition->last_stable_offset()

- The purpose of this change is to explicity use the cluster layer
abstractions when reading from input topics so that coproc never ingests
uncommitted data. That is because reading uncomitted data will
eventually produce uncommitted data to materialized topics, which would
then need to be truncated in the event a source topic is truncated.
Reading up until the last stable offset ensures we will never need to
handle truncations of the type that delete the end of the log.

- Fixes: #1623

- Fixes: #1596