storage: use a fixed workspace allocation for zstd
Generally speaking when using stream decompression a window will be
required whose size is determiend by parameters of the compression
process. It looks like the default comrpession settings in something
like librdkafka result in a window size need of 2 MB. This gets larger
the more extreme paramters are used during compression.

In this PR we allocate a fixed size workspace at early bootup when
fragmentation isn't an issue and then re-use this for the lifetime of
the process.

This ticket is tracking future improvements:

  https://github.com/vectorizedio/redpanda/issues/1802

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>