config: add control over zstd decompression buffer
The decompression buffer is a chunk of memory allocated per core
immediately after redpanda starts (when lots of large contiguous regions
of memory are available for the taking). The size of this window is
dependent on producers that send compressed data. Default settings for
clients I've seen result in 2 MBs being needed. Defaulting to 8mb. I
think principle the size is unbounded, but when using pre-set
compression limits it seems that we reach 8 MB when --ultra compression
is used. However, this is unlikely to be used since kafka clients are
generally used to batch size limits of a few MBs at most.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>