storage: make snapshot manager support multi snapshots
Snapshot manager used to work only with one target file so when we
needed to create several snapshots we had to create several snapshot
managers.

This commit introduces multi_snapshot_manager which works just like
snapshot_manager but instead of specifying the name of the snapshot
once in the constructor it let to specify it for each read & write
methods.