s/segment: fixed moving committed offset backward
Recent changes in segment appender made it possible to have multiple
concurrent appender flush operations pending. This lead to a race
condition in which flush with was dispatched earlier in terms of log
offset was finished after later one and caused committed offset to move
backward. This led to the situation in which caller after storing offset
and then executed `log::flush` operation observed committed offset
smaller than the one which was previously appended to the log.

Example:
```
[fiber 1] append -> received offset 10
[fiber 2] append -> received offset 12
[fiber 2] dispatch flush
[fiber 1] dispatch flush
[fiber 2] flush and update committed offset to 12
[fiber 1] flush and update committed offset to 10
[fiber 2] check if committed offset is >= 12 <--- inconsistency
```
Signed-off-by: Michal Maslanka <michal@vectorized.io>