kafka/tx: fix aborted txns related resource leak
Redpanda needs to store a list of all ever aborted transactions.

We used to do by reading the log and then keeping the list in a snapshot.
But over time this leads to a resource leak since the list of aborted
transactions just keeps growing.

Now we also reconstruct a list of the aborted transactions by reading
the log but once its size passes a threshold (abort_index_segment_size)
we save the list to a separate file (abort segment) in the log's work
directory and reset the list.

The file name includes first and the last offsets of the stored aborted
transactions e.g. "abort.idx.50000.100000".

Once an abort segment is saved we include a range it covers (e.g. 50000-
-100000) into an index stored in the snapshot. We use the index when we
serve fetch requests to look up a list of the aborted transactions by a
given range: first we look for the segments intersecting with the range,
then we load the picked segments and do more detailed filtration.

We assume that the most reads happen at the end of the log we cache the
last segment in memory to avoid the disk IO. As a result for each log
we store at most 2x abort_index_segment_size aborted transactions in
memory.