raft: log leader transfer errors at warning level
This is to aid with debugging of the new feature. The expectation is
that this will also not be unnecessarily noisey: we expect healthy raft
groups to have no trouble transferring leadership. And automatic
rebalancing attempts to throttle attempts with groups that appear
unhealthy.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit c5d50022c8a199701c3e5ea2d4c93a1c53dfb078)