rpk/config: Change format of rpk.admin_api.addresses
In the previous changeset that introduced the `rpk.kafka_api` and
`rpk.admin_api` config fields, it was agreed that the lists of
addresses should be represented as lists of strings, instead of
objects with separate "address" and "port" fields, since it's
easier for the users and there's no internal technical requirement
or restriction that implies it. `rpk.kafka_api.brokers` was changed
but `rpk.admin_api.addresses` was not.