admin/acl & cli/cmd/acl: remove mock, export type
This converts the admin package to the principle of "accept interfaces,
return structs". This is written about all over, as a first link:

  https://medium.com/@cep21/what-accept-interfaces-return-structs-means-in-go-2fe879e25ee8

which also links to Dave Cheney's post including this topic as well.

The benefit here is that now we can add blindly add functions to the
admin API without having the _requirement_ of updating our mock. A mock
also makes any new function technically a breaking change, meaning that
any mock implementations in different packages break.

Instead, we move the current usage for UserACLs into the one place the
mock is needed, the acl_test.

This commit will make it easier to add other functions to the admin
interface.