redpanda admin command: create, and support /v1/brokers
The redpanda admin command will talk to redpanda admin hosts. The
command space is long, but with tab completion this should be of no
concern, as well, admin commands are more advanced and we do not want
extremely high discoverability here.

The intent is to support all admin request paths.

To start, we support the brokers request paths, which now allows for
decommissioning (and potentially recommissioning) a broker through rpk.