s3: Do not cache signature key
The signature key is valid for two days so the implementation is
caching it for performance reasons.

This solution led to some bugs when persistent connections are
involved. When the key becomes stale we're getting a lot of S3
errors with error code 'AuthorizationHeaderMalformed' on AWS and
'MalformedSecurtyHeader' on GCS.

The bugs are particulariliy dangerous since the log uploads
are not possible until the signature key is updated.
The botocore library doesn't cache the signature key. We can assume
that it's perfectly fine for us to not to cache it as well. This commit
removes this logic.