raft: add ensure_disconnect() method to rpc_client_protocol
This provides a way to get at the underlying transport
object to close its socket.  This enables anyone holding such
a client to ensure that on the next use of the client, reconnect_transport
will try to reconnect.

In general, this function should not be used: the networking stack should
close dead connections, and reconnect_transport handles re-opening fresh
connections.  However, in some cases the network stack fails to tear
down a connection, and in those cases the heartbeat_manager can use
this method to force it.

Signed-off-by: John Spray <jcs@vectorized.io>