http: Add timeout for persistent https connections
AWS S3 uses 5s timeout for all https connections. S3 connection
pool in redpanda didn't use any timeouts before this commit.
Because of that, when the connection wasn't used for more than 5s
and was closed by the server redpanda still treated it as opened.
When such connection is being used for the next time it receives
a "broken pipe" error which means that the underlying TCP connection
was closed by the peer. Such error is treated by archival service
as transient and expected failure that can be retried immediately.
So no harm was done by this. But..

The error message is logged in RPC layer and looks like a bug in the
logs. The logging is not controlled by the http package so it's not
possible to disable it if we know that it's expected.

Also, sometimes in this case seastar just returns an empty buffer
instead of error code. The empty buffer means that connection is closed
gracefully by peer. But in this case we don't receive any response
bytes which triggers boost::beast short_read error. Same error is
usually connected with throttling. It appears when we're trying to
use connection afeter receiving "SlowDown" response. So seeing a
short_read error in this case is very misleading and can mask actual
throttling events.

This commit fixes this by adding a timestamp field to the http client.
The field is updated every time we receive something from the server.
When the new request needs to be made we're checking this fields and
if the timestamp is older than 5s the connection needs to be reset.

The default idle connection timeout here is set to 1s but it's only
used for tests.