cluster: Refuse join_request if rpc_address conflicts with existing one
This can happen if an operator makes a mistake in their configs.  Prior
to this change, a new node could be accepted into the cluster with
a bogus RPC address, with confusing results.

For example, if a new node has rpc address 0.0.0.0:33145 in its config,
then the leader would accept it, and then start trying to send AppendEntries
RPCs to that address... which would lead to the leader connecting to itself.

Signed-off-by: John Spray <jcs@vectorized.io>