kafka/tx: fix a try_abort & begin_tx deadlock
tx coordinator commits a transaction in two phases. once it passed
the first phase a data partition can't decide to abort the (expired)
transaction locally and needs to communicate with the coordinator
via try_abort RPC call. On the receiver part it takes a lock by tx id
and on the sender part it takes a lock by a producer id.

A fault may cause a coordinator to lose its latest state and to forget
about an ongoing transaction. A recreated producer may reuse the same
producer id (with bumped epoch) and the same tx id. In this case
try_abort may deadlock with begin_tx.

try_abort doesn't need a lock to detect this situation: existence of
the concurrent transaction implies the original is lost and it's safe
to abort. Fixing it and adding timeout to the lock as a safety net.