kafka/tx: fix stalls on truncation
Before log_eviction_stm confirms prefix truncation it makes sure that
the state machines are caught up with the log and have an actual snapshot
(ensure_snapshot_exists). The latter depends on state_machine::wait which
depends on the apply upcalls which stalls for acks=1.

comitted_index controls the apply upcalls. The fix updates comitted_index
to reflect the replicated front of the log.