admin api: add sendOne, sendAny
- adds sendOne to send a request to *one* url. The function expects that
the client will only have one url. This allows commands to select one
host, create the admin client with that single host, and use commands
that should only go to one host.

- adds sendAny to send a request to a randomly chosen broker. This
basically completes the ways one could send; sendAny is meant to be used
for commands that can be issued to any broker.