add package out
The out package is to be used as a quick way to print tab formatted
output and exit, as well as check errors and exit. This is basically a
helper to simplify extremely common 3 liners. As well, cobra's RunE does
not really add any context to errors, so simply returning errors is 99%
of the time not too informative. Using out allows for more obviously
annotating errors and exiting.