root: move listPlugins to a dedicated package
Previously, all plugin logic was contained in the root file. We now move
some of that to a dedicated plugin package -- particularly, we move the
plugin loading, as well as the struct definition and a few helper
functions.

We will need to list plugins in our incoming plugin command, so it is
better to have all of this logic in one place. We now remove the debug
logs from listing, they were not really relevant. We also now return
shadowed paths (which were the most important debug logs). The shadowed
paths will be useful for the plugin listing output.

All plugin logic that operated on cobra.Command's remains in root.go.