storage: use fragmented vector for index state
Normal 1GB segments can lead to 128kb vectors in the segment index
state. But we've seen 256k and 512kb in core files. And this is without
compaction. When compaction is enabled and segments are combined, then
if the size is not reduced the resulting index size would be expected to
be much larger. This is problematic when memory is fragmented because we
may not be able to find contiguous memory regions.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit f27cc56b4d594d8447222eaf05aa7ca02cc79e84)