raft/heartbeat_manager: close connections if heartbeats are failing
This combines the state in consensus with the ensure_disconnect hook
in consensus_client_protocol.  The resulting behaviour fixes
an issue where Redpanda nodes can fail to reach consensus if
a node restarts and other nodes have 'zombie' TCP connections
to the previous instance of the node, by forcing nodes to open
fresh connections if too many heartbeats have failed.

Fixes: https://github.com/vectorizedio/support/issues/79

Signed-off-by: John Spray <jcs@vectorized.io>